// src/auth/enums/userStatus.enum.ts
export enum UserStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  BLOCKED = 'BLOCKED',
  DELETED = 'DELETED',
  NOT_COMPLETED = 'NOT_COMPLETED',
}

export enum UserStatusForVendorReview {
  NOT_COMPLETED = 'NOT_COMPLETED',
  ACTIVE = 'ACTIVE',
}

// delete only
export enum UserDELETEStatus {
  DELETED = 'DELETED',
}